refactor: move the wallet exports and shared marshalling into zcash_jni
Move the fifty `RustBackend` exports into `zcash_jni::wallet`, and the
marshalling that more than one area needs into `zcash_jni` itself. This is
the commit that empties lib.rs: 2711 lines to 152, and 50 exports to none.

Still a pure move. Every relocated function keeps the body it had at
d65b4d2c, `wallet_db` and `block_db` included: they arrive here whole, in the
form they had in lib.rs. Splitting them is a real edit, so it is done in the
next two commits where the diff is small enough to read.

The shared helpers had to move in the same commit rather than a later one.
They were defined in the middle of the wallet export region, so lifting the
exports out of lib.rs took `decode_usk` and `path_from_jni` with them, and
`zcash_jni::derivation` and `zcash_jni::tor` already called both.

Placement follows the rule that marshalling lives beside the exports it
serves. `zcash_jni/mod.rs` holds only what two or more submodules use:
`path_from_jni`, `secret_from_jni`, `account_id_from_jni`, `parse_txid`,
`parse_treestate`, `parse_pczt`, `parse_ufvk`, `parse_ua`,
`zip32_account_index_from_jlong`, `decode_usk`, and the two database
openers. `encode_usk` stayed in `wallet.rs` and `encode_transaction` moved
to `tor.rs`, each having exactly one caller.

Nothing needed widening to `pub(crate)`. A private item is already in scope
for descendant modules, and every helper moved either from the crate root
or into an ancestor of its callers.

lib.rs is now logic only: `ReceiverFlags`, `zip317_helper`, four `parse_*`
conversions over plain integers, and `UnifiedAddressParser`. It no longer
names the `jni` crate at all, which is the invariant this refactor is for.
`crate::utils` still does, by design: it is the domain-neutral JNI plumbing
every submodule builds on, not a logic module.

Exported symbol set unchanged: 78 before, 78 after, empty `nm -gU` diff.

    ./scripts/check-jni-symbols.sh          # 78 symbols, baseline matches
    ./scripts/check-jni-move.sh d65b4d2c    # 78 byte-identical, 0 changed
